Generalize Scores.
