+++ src/main/java/io/jenkins/plugins/grading/AnalysisScore.java
@@ -1,26 +1,20 @@
 package io.jenkins.plugins.grading;
 
-import edu.hm.hafner.util.Ensure;
-
 import io.jenkins.plugins.analysis.core.model.AnalysisResult;
 
 /**
- * Computes the {@link Score} impact of static analysis results. These results are obtained by inspecting a {@link
+ * Computes the {@link AggregatedScore} impact of static analysis results. These results are obtained by inspecting a {@link
  * AnalysisResult} instance of the Warnings Next Generation plugin.
  *
  * @author Eva-Maria Zeintl
  */
 @SuppressWarnings("PMD.DataClass")
-public class AnalysisScore {
-    private final String id;
-    private final String name;
-
-    private final int totalImpact;
-
+public class AnalysisScore extends Score {
     private final int errorsSize;
     private final int highPrioritySize;
     private final int normalPrioritySize;
     private final int lowPrioritySize;
+
     private final int totalSize;
 
     /**
@@ -34,11 +28,7 @@ public class AnalysisScore {
      *         the static analysis result
      */
     public AnalysisScore(final String name, final AnalysisConfiguration configuration, final AnalysisResult result) {
-        Ensure.that(name).isNotEmpty();
-        Ensure.that(result.getId()).isNotEmpty();
-
-        this.name = name;
-        this.id = result.getId();
+        super(result.getId(), name);
 
         this.errorsSize = result.getTotalErrorsSize();
         this.highPrioritySize = result.getTotalHighPrioritySize();
@@ -46,7 +36,7 @@ public class AnalysisScore {
         this.lowPrioritySize = result.getTotalLowPrioritySize();
         this.totalSize = result.getTotalSize();
 
-        totalImpact = computeImpact(configuration);
+        setTotalImpact(computeImpact(configuration));
     }
 
     private int computeImpact(final AnalysisConfiguration configuration) {
@@ -58,18 +48,6 @@ public class AnalysisScore {
         change = change + configuration.getLowImpact() * lowPrioritySize;
 
         return change;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getTotalImpact() {
-        return totalImpact;
     }
 
     public int getErrorsSize() {
