+++ src/main/java/io/jenkins/plugins/grading/ScoreXmlStream.java
@@ -1,29 +1,30 @@
 package io.jenkins.plugins.grading;
 
 import hudson.util.XStream2;
+
 import io.jenkins.plugins.util.AbstractXmlStream;
 
 /**
- * Reads {@link Score} from an XML file.
+ * Reads {@link AggregatedScore} from an XML file.
  *
  * @author Eva-Maria Zeintl
  */
-public class ScoreXmlStream extends AbstractXmlStream<Score> {
+public class ScoreXmlStream extends AbstractXmlStream<AggregatedScore> {
     /**
      * creates a new {@link ScoreXmlStream}.
      */
     public ScoreXmlStream() {
-        super(Score.class);
+        super(AggregatedScore.class);
     }
 
     @Override
-    protected Score createDefaultValue() {
-        return new Score();
+    protected AggregatedScore createDefaultValue() {
+        return new AggregatedScore();
     }
 
     @Override
     protected void configureXStream(final XStream2 xStream) {
-        xStream.alias("scores", Score.class);
+        xStream.alias("scores", AggregatedScore.class);
     }
 
 }
