+++ src/test/java/io/jenkins/plugins/grading/AutoGraderITest.java
@@ -8,7 +8,6 @@ import org.jvnet.hudson.test.JenkinsRule;
 
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
-
 import hudson.model.Result;
 import hudson.model.Run;
 
@@ -65,7 +64,7 @@ public class AutoGraderITest extends IntegrationTestWithJenkinsPerSuite {
 
         List<AutoGradingBuildAction> actions = baseline.getActions(AutoGradingBuildAction.class);
         assertThat(actions).hasSize(1);
-        Score score = actions.get(0).getResult();
+        AggregatedScore score = actions.get(0).getResult();
 
         assertThat(score).hasAchieved(40);
     }
