+++ src/test/java/io/jenkins/plugins/grading/ScoreTest.java
@@ -14,7 +14,7 @@ import static io.jenkins.plugins.grading.assertions.Assertions.*;
 import static org.mockito.Mockito.*;
 
 /**
- * Tests the class {@link Score}.
+ * Tests the class {@link AggregatedScore}.
  *
  * @author Ullrich Hafner
  * @author Oliver Scholz
@@ -22,7 +22,7 @@ import static org.mockito.Mockito.*;
 class ScoreTest {
     @Test
     void shouldSumAnalysisConfiguration() {
-        Score score = new Score();
+        AggregatedScore score = new AggregatedScore();
 
         assertThat(score).hasAchieved(0);
 
@@ -31,19 +31,19 @@ class ScoreTest {
         assertThat(score).hasAchieved(20);
         assertThat(score).hasTotal(20);
         assertThat(score).hasRatio(100);
-        assertThat(score).hasStyle(Score.EXCELLENT);
+        assertThat(score).hasStyle(AggregatedScore.EXCELLENT);
 
         score.addAnalysisTotal(configuration, Collections.singletonList(createAnalysisScore(-10)));
         assertThat(score).hasAchieved(30);
         assertThat(score).hasTotal(40);
         assertThat(score).hasRatio(75);
-        assertThat(score).hasStyle(Score.EXCELLENT);
+        assertThat(score).hasStyle(AggregatedScore.EXCELLENT);
 
         score.addAnalysisTotal(configuration, Arrays.asList(createAnalysisScore(-10), createAnalysisScore(-5)));
         assertThat(score).hasAchieved(35);
         assertThat(score).hasTotal(60);
         assertThat(score).hasRatio(35 * 100 / 60);
-        assertThat(score).hasStyle(Score.GOOD);
+        assertThat(score).hasStyle(AggregatedScore.GOOD);
     }
 
     private AnalysisScore createAnalysisScore(final int total) {
@@ -61,7 +61,7 @@ class ScoreTest {
                 .setMissedImpact(-1)
                 .build();
 
-        Score score = new Score();
+        AggregatedScore score = new AggregatedScore();
         score.addCoverageTotal(coverageConfiguration,
                 new CoverageScore("Line", coverageConfiguration, Ratio.create(50, 100)),
                 new CoverageScore("Branch", coverageConfiguration, Ratio.create(60, 100)));
@@ -83,7 +83,7 @@ class ScoreTest {
         TestScore testScore = mock(TestScore.class);
         when(testScore.getTotalImpact()).thenReturn(-3);
 
-        Score score = new Score();
+        AggregatedScore score = new AggregatedScore();
         score.addTestsTotal(testConfiguration, testScore);
 
         assertThat(score).hasAchieved(97);
@@ -103,7 +103,7 @@ class ScoreTest {
         PitScore pitScore = mock(PitScore.class);
         when(pitScore.getTotalImpact()).thenReturn(-2);
 
-        Score score = new Score();
+        AggregatedScore score = new AggregatedScore();
         score.addPitTotal(pitConfiguration, pitScore);
 
         assertThat(score).hasAchieved(98);
@@ -115,7 +115,7 @@ class ScoreTest {
 
     @Test
     void shouldUpdateWithAllConfigurations() {
-        Score score = new Score();
+        AggregatedScore score = new AggregatedScore();
         assertThat(score).hasAchieved(0);
 
         PitConfiguration pitConfiguration = new PitConfiguration.PitConfigurationBuilder()
@@ -130,7 +130,7 @@ class ScoreTest {
         assertThat(score).hasAchieved(80);
         assertThat(score).hasTotal(100);
         assertThat(score).hasRatio(80);
-        assertThat(score).hasStyle(Score.EXCELLENT);
+        assertThat(score).hasStyle(AggregatedScore.EXCELLENT);
         assertThat(score.getPitConfiguration()).isSameAs(pitConfiguration);
         assertThat(score).hasPitScores(pitScore);
 
@@ -146,7 +146,7 @@ class ScoreTest {
         assertThat(score).hasAchieved(120);
         assertThat(score).hasTotal(200);
         assertThat(score).hasRatio(60);
-        assertThat(score).hasStyle(Score.GOOD);
+        assertThat(score).hasStyle(AggregatedScore.GOOD);
         assertThat(score.getTestConfiguration()).isSameAs(testConfiguration);
         assertThat(score).hasTestScores(testScore);
 
@@ -162,7 +162,7 @@ class ScoreTest {
         assertThat(score).hasAchieved(150);
         assertThat(score).hasTotal(300);
         assertThat(score).hasRatio(50);
-        assertThat(score).hasStyle(Score.GOOD);
+        assertThat(score).hasStyle(AggregatedScore.GOOD);
         assertThat(score.getCoverageConfiguration()).isSameAs(coverageConfiguration);
         assertThat(score).hasCoverageScores(coverageScore);
 
@@ -176,7 +176,7 @@ class ScoreTest {
         assertThat(score).hasAchieved(199);
         assertThat(score).hasTotal(400);
         assertThat(score).hasRatio(49);
-        assertThat(score).hasStyle(Score.FAILURE);
+        assertThat(score).hasStyle(AggregatedScore.FAILURE);
         assertThat(score.getAnalysisConfiguration()).isSameAs(analysisConfiguration);
         assertThat(score.getAnalysisScores()).containsExactly(analysisScore);
     }
